Have changed to usage of class and inherited PubSub. The class is working when creating an RTDServer object.
Todo includes, create a method which is called and assigned to variable hd_str inside some_historical_data. Ensure correct parsing of data, as all the hard work was already done by ideepcoder.

Use CCXT for crypto exchaneg data. We'll need to create exchange instances and call await exchange.close() when we are finished with exchange instances to prevent errors. Create a naming convention based on the exchange, e.g BTC/USDT.bin would be Bitcoin in USDT on Binance, and BTC/USDT.kra would be the same on Kraken etc. The is proabably best achieved either by use of a dictionary, or by match case statement. Implement this using methods and return values and we should be able to implement multiple exchange instances at once. Assign exchange instances to a list so that when we are exiting, we can call a method and iterate though, e.g for exchange in exchange_list: await exchange.close().

This is going to be so much fun!!