Special characters do not seem to be transmitted therefore cannot be used in symbol names here. So we'll just name the symbol and append it with the exchange name and just parse it as needed. To make things a bit neater, when parsing on the python side and sending into CCXT, symbols could be .upper(), however, this is probably not necessary in most cases as ccxt abstracted away most of this stuff and has unified pairs in most cases. Address it if it is a problem.
This first commit is working ok for integrating fetch_ohclv(), but in order to speed it up, we'll probably append messages to a queue, read and broadcast them from there, in it's own loop, rather than call/ response.
This will be important when calling multiple symbols asynchronously.

The quotes are populated from the script, but because the RTD is not yet up and running, we are getting lower values distorting out charts. So next task is to sore the RTD websocket stuff.

Then later we'll want to address all of the different possible outcomes from the plug-in call to the server (in recv), and we'll also want to create a method to unify the timeframe of the call to ccxt, with the timeframe of the database (if we want higher timeframes, to enable more data sharing.

It would be good to clarify also, that, when we do a first data call, we want to call on a limit of 1000. But, it would be useful to perform some calculation (to be able to set 'since' parameter to the exchange) and then paginate 1w, 1d, 4h, 1h, 15m 5m,1m. 1w should be first. Also, user should specify allow mixed and EOD data on Database Creation.